Ns descriptor related stuffs.
Modify the biz for uploading of nsd content.

Issue-ID: VFC-1037

+++ catalog/packages/biz/ns_descriptor.py
@@ -155,21 +155,23 @@ def delete_single(nsd_info_id):
     ns_pkgs.delete()
 
 
-def upload(files, nsd_info_id):
-    remote_files = files
-    for remote_file in remote_files:
-        local_file_name = remote_file.name
-        local_file_dir = os.path.join(CATALOG_ROOT_PATH, nsd_info_id)
-        local_file_name = os.path.join(local_file_dir, local_file_name)
-        if not os.path.exists(local_file_dir):
-            fileutil.make_dirs(local_file_dir)
-        with open(local_file_name, 'wb') as local_file:
-            if remote_file.multiple_chunks(chunk_size=None):
-                for chunk in remote_file.chunks():
-                    local_file.write(chunk)
-            else:
-                data = remote_file.read()
-                local_file.write(data)
+def upload(remote_file, nsd_info_id):
+    ns_pkgs = NSPackageModel.objects.filter(nsPackageId=nsd_info_id)
+    if not ns_pkgs.exists():
+        raise CatalogException('The NS descriptor (%s) does not exist.' % nsd_info_id)
+
+    local_file_name = remote_file.name
+    local_file_dir = os.path.join(CATALOG_ROOT_PATH, nsd_info_id)
+    local_file_name = os.path.join(local_file_dir, local_file_name)
+    if not os.path.exists(local_file_dir):
+        fileutil.make_dirs(local_file_dir, 0o777)
+    with open(local_file_name, 'wb') as local_file:
+        if remote_file.multiple_chunks(chunk_size=None):
+            for chunk in remote_file.chunks():
+                local_file.write(chunk)
+        else:
+            data = remote_file.read()
+            local_file.write(data)
 
 
 def fill_resp_data(ns_pkg):
